requirements.txt
# Core dependencies
pandas>=2.1.0
numpy>=1.24.0
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU is available

# Web scraping
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Web interface
streamlit>=1.29.0
plotly>=5.18.0

# Image processing
Pillow>=10.0.0

# Utilitiespip install -r requirements.txt
tqdm>=4.66.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Optional: GPU support
# cupy-cuda11x  # Uncomment and install appropriate version for your CUDA toolkit
# faiss-gpu  # Uncomment for GPU support
